response codes 200, 404, etc

+++ src/main/java/com/example/demo/BooksController.java
@@ -5,9 +5,13 @@ import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.List;
+import java.util.Optional;
 import java.util.Scanner;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -24,27 +28,77 @@ public class BooksController {
 
 	// @RequestMapping(value = "/books", method = RequestMethod.GET)
 	@GetMapping(value = "/books")
-	public List<Book> getBooks() {
-		return this.bookService.getBooks();
+	public ResponseEntity<List<Book>> getBooks() {
+		List<Book> bookList = this.bookService.getBooks();
+		try {
+			if (bookList.size() > 0) {
+				return ResponseEntity.of(Optional.of(bookList));
+			} else {
+				return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+		}
+
 	}
 
 	@GetMapping(value = "/books/{id}")
-	public Book getBookById(@PathVariable("id") int id) {
-		return this.bookService.getBookByID(id);
+	public ResponseEntity<Book> getBookById(@PathVariable("id") int id) {
+		Book book = this.bookService.getBookByID(id);
+
+		try {
+			if (book != null) {
+				return ResponseEntity.of(Optional.of(book));
+			} else {
+				return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+		}
 	}
-	
+
 	@PostMapping("/books")
-	public Book addBook(@RequestBody Book book) {
-		return this.bookService.addBook(book);
+	public ResponseEntity<Void> addBook(@RequestBody Book book) {
+		try {
+			this.bookService.addBook(book);
+			return ResponseEntity.status(HttpStatus.OK).build();
+		} catch (Exception e) {
+			e.printStackTrace();
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+		}
 	}
-	
+
 	@DeleteMapping(value = "/books/{id}")
-	public Book deleteBookById(@PathVariable("id") int id) {
-		return this.bookService.deleteBookByID(id);
+	public ResponseEntity<Book> deleteBookById(@PathVariable("id") int id) {
+		try {
+			Book book = this.bookService.getBookByID(id);
+			if (book != null) {
+				this.bookService.deleteBookByID(id);
+				return ResponseEntity.status(HttpStatus.OK).build();
+			} else {
+				return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+		}
 	}
-	
+
 	@PutMapping("/books/{id}")
-	public Book updateBook(@RequestBody Book book, @PathVariable("id") int id) {
-		return this.bookService.updateBook(book, id);
+	public ResponseEntity<Book> updateBook(@RequestBody Book book, @PathVariable("id") int id) {
+		try {
+			Book _book = this.bookService.getBookByID(id);
+			if (_book != null) {
+				this.bookService.updateBook(book, id);
+				return ResponseEntity.status(HttpStatus.OK).build();
+			} else {
+				return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+		}
 	}
 }
